Removed old HTTP plugins/services, the new one is now named "http".
